# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Invest Agents Web UI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Login Api Auth Login Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/portfolio/dime/emails": {
      "get": {
        "tags": [
          "Dime Sync"
        ],
        "summary": "Search Dime Emails Endpoint",
        "operationId": "search_dime_emails_endpoint_api_portfolio_dime_emails_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Query"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DimeEmailListResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/dime/scan/email": {
      "post": {
        "tags": [
          "Dime Sync"
        ],
        "summary": "Scan Email Attachment Endpoint",
        "operationId": "scan_email_attachment_endpoint_api_portfolio_dime_scan_email_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DimeScanEmailRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DimeScanResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/dime/scan/upload": {
      "post": {
        "tags": [
          "Dime Sync"
        ],
        "summary": "Scan Upload Pdf Endpoint",
        "operationId": "scan_upload_pdf_endpoint_api_portfolio_dime_scan_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_scan_upload_pdf_endpoint_api_portfolio_dime_scan_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DimeScanResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/dime/staged/{scan_id}": {
      "get": {
        "tags": [
          "Dime Sync"
        ],
        "summary": "Get Staged Endpoint",
        "operationId": "get_staged_endpoint_api_portfolio_dime_staged__scan_id__get",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DimeScanResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/dime/commit/{scan_id}": {
      "post": {
        "tags": [
          "Dime Sync"
        ],
        "summary": "Commit Staged Endpoint",
        "operationId": "commit_staged_endpoint_api_portfolio_dime_commit__scan_id__post",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DimeCommitRequestDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DimeCommitResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActualPortfolioStateDTO": {
        "properties": {
          "last_updated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Updated"
          },
          "fx_rates": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Fx Rates",
            "default": {}
          },
          "summary": {
            "$ref": "#/components/schemas/ActualSummaryDTO"
          },
          "allocation_targets": {
            "items": {
              "$ref": "#/components/schemas/AllocationTargetDTO"
            },
            "type": "array",
            "title": "Allocation Targets",
            "default": []
          },
          "holdings": {
            "items": {
              "$ref": "#/components/schemas/ActualHoldingDTO"
            },
            "type": "array",
            "title": "Holdings",
            "default": []
          },
          "price_refresh_info": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Refresh Info"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "ActualPortfolioStateDTO"
      },
      "ActualSummaryDTO": {
        "properties": {
          "total_value_thb": {
            "type": "number",
            "title": "Total Value Thb",
            "default": 0.0
          },
          "total_cost_basis_thb": {
            "type": "number",
            "title": "Total Cost Basis Thb",
            "default": 0.0
          },
          "total_unrealized_profit": {
            "type": "number",
            "title": "Total Unrealized Profit",
            "default": 0.0
          },
          "total_realized_profit_ytd": {
            "type": "number",
            "title": "Total Realized Profit Ytd",
            "default": 0.0
          },
          "passive_income_ytd": {
            "type": "number",
            "title": "Passive Income Ytd",
            "default": 0.0
          },
          "total_accumulated_dividend": {
            "type": "number",
            "title": "Total Accumulated Dividend",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "ActualSummaryDTO"
      },
...
      "Body_scan_upload_pdf_endpoint_api_portfolio_dime_scan_upload_post": {
        "properties": {
          "pdf_file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Pdf File"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "pdf_file"
        ],
        "title": "Body_scan_upload_pdf_endpoint_api_portfolio_dime_scan_upload_post"
      },
...
      "DimeCommitRequestDTO": {
        "properties": {
          "portfolio_id": {
            "type": "string",
            "title": "Portfolio Id",
            "default": "default"
          }
        },
        "type": "object",
        "title": "DimeCommitRequestDTO"
      },
      "DimeCommitResponseDTO": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "imported_count": {
            "type": "integer",
            "title": "Imported Count"
          },
          "state": {
            "$ref": "#/components/schemas/ActualPortfolioStateDTO"
          }
        },
        "type": "object",
        "required": [
          "imported_count",
          "state"
        ],
        "title": "DimeCommitResponseDTO"
      },
      "DimeEmailListResponseDTO": {
        "properties": {
          "emails": {
            "items": {
              "$ref": "#/components/schemas/DimeEmailMetadataDTO"
            },
            "type": "array",
            "title": "Emails",
            "default": []
          }
        },
        "type": "object",
        "title": "DimeEmailListResponseDTO"
      },
      "DimeEmailMetadataDTO": {
        "properties": {
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "attachment_id": {
            "type": "string",
            "title": "Attachment Id"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "sender": {
            "type": "string",
            "title": "Sender"
          },
          "received_at": {
            "type": "string",
            "title": "Received At"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          }
        },
        "type": "object",
        "required": [
          "message_id",
          "attachment_id",
          "subject",
          "sender",
          "received_at",
          "filename",
          "size_bytes"
        ],
        "title": "DimeEmailMetadataDTO"
      },
      "DimeScanEmailRequestDTO": {
        "properties": {
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "attachment_id": {
            "type": "string",
            "title": "Attachment Id"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "message_id",
          "attachment_id"
        ],
        "title": "DimeScanEmailRequestDTO"
      },
      "DimeScanResponseDTO": {
        "properties": {
          "scan_id": {
            "type": "string",
            "title": "Scan Id"
          },
          "item_count": {
            "type": "integer",
            "title": "Item Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DimeStagedItemDTO"
            },
            "type": "array",
            "title": "Items",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "scan_id",
          "item_count"
        ],
        "title": "DimeScanResponseDTO"
      },
      "DimeStagedItemDTO": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "trade_date": {
            "type": "string",
            "title": "Trade Date"
          },
          "settlement_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settlement Date"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "units": {
            "type": "string",
            "title": "Units"
          },
          "price": {
            "type": "string",
            "title": "Price"
          },
          "gross_amount": {
            "type": "string",
            "title": "Gross Amount"
          },
          "fees": {
            "$ref": "#/components/schemas/DimeStagedItemFeeDTO"
          },
          "net_amount": {
            "type": "string",
            "title": "Net Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "THB"
          },
          "exchange_rate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange Rate"
          },
          "confirmation_no": {
            "type": "string",
            "title": "Confirmation No"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "DIME"
          },
          "fingerprint": {
            "type": "string",
            "title": "Fingerprint"
          },
          "line_index": {
            "type": "integer",
            "title": "Line Index",
            "default": 0
          },
          "cash_adjusted": {
            "type": "boolean",
            "title": "Cash Adjusted",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "trade_date",
          "symbol",
          "action",
          "units",
          "price",
          "gross_amount",
          "net_amount",
          "confirmation_no",
          "fingerprint"
        ],
        "title": "DimeStagedItemDTO"
      },
      "DimeStagedItemFeeDTO": {
        "properties": {
          "commission": {
            "type": "string",
            "title": "Commission",
            "default": "0.00"
          },
          "vat": {
            "type": "string",
            "title": "Vat",
            "default": "0.00"
          },
          "other_fees": {
            "type": "string",
            "title": "Other Fees",
            "default": "0.00"
          },
          "fee_currency": {
            "type": "string",
            "title": "Fee Currency",
            "default": "THB"
          }
        },
        "type": "object",
        "title": "DimeStagedItemFeeDTO"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}